# Feature test click_dropdown performance