--- a/spec/support/feature_helper.rb
+++ b/spec/support/feature_helper.rb
@@ -132,7 +132,7 @@ module FeatureHelper
   end
 
   def dropdown_menu_visible?(dropdown)
-    dropdown.sibling(".cf-select__menu")
+    dropdown.sibling(".cf-select__menu", wait: false)
   rescue Capybara::ElementNotFound
     false
   else