… spec/feature/hearings/schedule_veteran/build_hearsched_spec.rb