Resolves 5+ second call we make 308 times in feature tests

### Description
Every time we call the spec feature helper `click_dropdown`, we call a [function](https://github.com/department-of-veterans-affairs/caseflow/blob/e5578264afe56c64744e45a02942f179c61450d0/spec/support/feature_helper.rb#L134-L140) to check to see if the actual menu is open. The call relies on the method `sibling` which waits 5 seconds before throwing an `Capybara::ElementNotFound`. This means every call to `click_dropdown` takes at least 5 seconds to run. We have 308+ instances of `click_dropdown` in our tests (+ as they're used in shared examples as well).
```bash
cd spec/feature
git grep click_dropdown | wc -l
308
```

Removing this wait time drops the call down to less than 1 second and resulted in no flakes when run in ci with an ensure_stable

### Acceptance Criteria
- [x] [No flakes](https://app.circleci.com/pipelines/github/department-of-veterans-affairs/caseflow/36193/workflows/5d26a78f-1b75-46d5-bdaa-9f8ce0c93044/jobs/134468) when wrapping test file with most instances of `click_dropdown` in an [`ensure_stable`](https://github.com/department-of-veterans-affairs/caseflow/pull/15532/commits/481afd4b681ac0fbad1a76391c48d6334189af5a#diff-241535ff479333e36fdb6f0397437c0c02fb56df87058f9e86c1361a90cc9e7bR6)
- [ ] Faster feature tests

### Testing Plan
1. Benchmark a single call! Throw a binding.pry [here](https://github.com/department-of-veterans-affairs/caseflow/blob/8ab2355f4be5472ca6b3125edc2c094fe236d77c/spec/feature/queue/mail_task_spec.rb#L41)
    1. master: 
    ```ruby
    Benchmark.measure { click_dropdown(prompt: prompt, text: text) }
    => #<Benchmark::Tms:0x00007f817bc00018
     @cstime=0.0,
     @cutime=0.0,
     @label="",
     @real=5.62093500001356,
     @stime=0.11391000000000062,
     @total=0.7551169999999985,
     @utime=0.6412069999999979>
    ```
    1. this branch:
    ```ruby
    Benchmark.measure { click_dropdown(prompt: prompt, text: text) }
    => #<Benchmark::Tms:0x00007faf4a0f2a48
     @cstime=0.0,
     @cutime=0.0,
     @label="",
     @real=0.8050880003720522,
     @stime=0.04735200000000006,
     @total=0.25439900000000293,
     @utime=0.20704700000000287>
    ```
1. Benchmark an entire feature test file
    1. master:
    
    ```bash
    bundle exec rake SPEC=spec/feature/hearings/schedule_veteran/build_hearsched_spec.rb
    ...
    Finished in 14 minutes 58 seconds (files took 6.15 seconds to load)
    ```
  
    1. this branch:
    ```bash
    bundle exec rake SPEC=spec/feature/hearings/schedule_veteran/build_hearsched_spec.rb
    ...
    Finished in 4 minutes 9.9 seconds (files took 6.45 seconds to load)
    ```
1. Benchmark the whole suite

Before
<img width="847" alt="Screen Shot 2020-10-30 at 9 06 12 AM" src="https://user-images.githubusercontent.com/45575454/97708754-ac66dd00-1a8f-11eb-8d22-a36cd80b6a86.png">
After
<img width="850" alt="Screen Shot 2020-10-30 at 9 05 09 AM" src="https://user-images.githubusercontent.com/45575454/97708758-acff7380-1a8f-11eb-8dbe-4592af3a0855.png">